requirements.txt
requests==2.31.0
pytest==7.4.0
prometheus-client==0.17.1
fastapi==0.103.1
uvicorn==0.23.2
kafka-python==2.1.5